Fork the GameEnginePropertyReader so it can also read properties for the lobby. Remove lobby arg parsing in prep to use the properties file reader.
