+++ lobby.properties
@@ -1,0 +1,2 @@
+port  = 3304
+
