--- src/main/java/games/strategy/engine/config/PropertyNotFoundException.java
@@ -1,10 +1,0 @@
-package games.strategy.engine.config;
-
-public class PropertyNotFoundException extends IllegalStateException {
-  private static final long serialVersionUID = -7834937010739816090L;
-
-  public PropertyNotFoundException(final GameEngineProperty property) {
-    super("Could not find property: " + property.toString() + ", in game engine configuration file: "
-        + GameEnginePropertyFileReader.getConfigFilePath());
-  }
-}
